Create unicode module in datafusion/functions/src/unicode and unicode_expressions feature flag, move char_length function
