+++ datafusion/functions/src/string/lower.rs
@@ -15,12 +15,16 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::string::common::{handle, utf8_to_str_type};
+use std::any::Any;
+
 use arrow::datatypes::DataType;
+
 use datafusion_common::Result;
 use datafusion_expr::ColumnarValue;
 use datafusion_expr::{ScalarUDFImpl, Signature, Volatility};
-use std::any::Any;
+
+use crate::string::common::handle;
+use crate::utils::utf8_to_str_type;
 
 #[derive(Debug)]
 pub(super) struct LowerFunc {
