+++ datafusion/sql/Cargo.toml
@@ -49,6 +49,7 @@ strum = { version = "0.26.1", features = ["derive"] }
 
 [dev-dependencies]
 ctor = { workspace = true }
+datafusion-functions = { workspace = true, default-features = true }
 env_logger = { workspace = true }
 paste = "^1.0"
 rstest = { workspace = true }
